bioapi/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
requests==2.31.0
numpy==1.24.3
scipy==1.11.4
biopython==1.81
rdkit==2023.9.1
python-multipart==0.0.6
aiofiles==23.2.0
